+++ packages/gasket-plugin-docs/lib/create.js
@@ -1,7 +1,7 @@
 /// <reference types="create-gasket-app"/>
 /// <reference types="@gasket/plugin-git" />
 
-const { name, version } = require('../package.json');
+const { name, version, devDependencies } = require('../package.json');
 
 const { DEFAULT_CONFIG } = require('./utils/constants');
 
@@ -17,9 +17,15 @@ module.exports = function create(gasket, {
   if (!useDocs) return;
 
   gitignore?.add(DEFAULT_CONFIG.outputDir, 'documentation');
-  gasketConfig.addPlugin('pluginDocs', name);
-  pkg.add('dependencies', {
-    [name]: `^${version}`
+  gasketConfig.addCommand('docs', {
+    dynamicPlugins: [
+      `${name}`,
+      '@gasket/plugin-metadata'
+    ]
+  });
+  pkg.add('devDependencies', {
+    [name]: `^${version}`,
+    '@gasket/plugin-metadata': devDependencies['@gasket/plugin-metadata']
   });
 
   const docsScript = typescript